Pass more accurate reset flags to the application.
Change the way the reset flags are passed to the application in
register r2. With this change, the application will be able to
distinguish between an external reset and a watchdog reset.
Previously, both were reported as a watchdog reset.

Also add workarounds to prevent gcc (4.7.2/4.8.2) from creating a
stack frame for main() that conflicts with the new reset flags
code, and would cause a problem for ATmega8 and ATmega32 devices.