Allow loading an OSCCAL value from PROGMEM
When the OSCCAL_PROGMEM variable is given at compiletime, a byte is
allocated in PROGMEM (flash) at a fixed location at the top of the flash
space (just below the version bytes, so the third byte from the top).

If this byte is overwritten with a calibration value (e.g. when
programming the bootloader), it will be automatically loaded on startup.
This allows improving the accuracy of the UART baudrate when running off
the internal oscillator, and prevents the actual program from having to
worry about calibration.

The default value of this byte is 0xff, which causes it to be ignored.

To fix the address of this variable, a dedicated .osccal section is used
for it, and the linker is passed some options to put it in the right
place (identical to what happens for the version variable).

Enabling this option adds 15 bytes on the atmega328p build.