Dynamically derive the chip clock rate, so we no longer need AVR_FREQ, or separate binaries for different clockrates.
Dynamically read chip signature, which should permit quie a bit of sharing of single binaries between multiple chips (as long as LED and UARTTX are the same.)

Now that we have a slower clockrate, the LED delay only needs 16bits.

Update Makefile help text, and the options saved in flash with BIGBOOT.

Document fuse requirements/issues.