Fix STK_SW_MAJOR and STK_SW_MINOR request responses
The major and minor version are stored at the end of program space,
but the protocol handler falsely uses the `optiboot_version` address
as an address to RAM, although the data is actually allocated in FLASH
and should be retrieved by a dedicated `lpm` instructions.

CAUTION: this change increase the binary size 22 bytes on avr-gcc 11.2.0
and 26 bytes on avr-gcc 5.4.0. Therefore, the team might want to
look at different solutions that do not occupy that much code space.
Nonetheless, there is a bug in the current implementation of
STK_SW_MAJOR STK_SW_MINOR that needs addressing.